+++ src/Forms/resources/fields/file-window.blade.php
@@ -1,25 +1,32 @@
 @php
-    /** @var \Thinktomorrow\AssetLibrary\Asset[] $files */
-    $files = $getValue($locale);
-    $count = count($files);
+    /** @var \Thinktomorrow\AssetLibrary\Asset[] $assets */
+    $assets = $getValue($locale);
+    $count = count($assets);
 @endphp
 
 <div class="flex flex-wrap -space-x-2">
-    @forelse ($files as $file)
+    @forelse ($assets as $asset)
         <div class="flex gap-4">
-            <a href="{{ $file->getUrl() }}" title="Bestand bekijken" target="_blank" rel="noopener" @class([
+            <a href="{{ $asset->getUrl() }}" title="Bestand bekijken" target="_blank" rel="noopener" @class([
                 'border-2 border-white rounded-lg' => $count > 1,
             ])>
                 <div class="flex items-center justify-center overflow-hidden rounded-lg w-14 h-14 shrink-0 bg-grey-100">
-                    @if($file->isImage())
-                        {{-- TODO: need previewUrl here --}}
+                    @if($asset->isImage())
                         <img
-                            src="{{ $file->getUrl('thumb') }}"
-                            alt="{{ $file->getFileName() }}"
+                            src="{{ $asset->getUrl('thumb') }}"
+                            alt="{{ $asset->getFileName() }}"
+                            class="object-contain w-full h-full"
+                        >
+                    @elseif($asset instanceof \Thinktomorrow\AssetLibrary\External\ExternalAssetContract)
+                        <img
+                            src="{{ $asset->getPreviewUrl('thumb') }}"
+                            alt="{{ $asset->getFileName() }}"
                             class="object-contain w-full h-full"
                         >
                     @else
-                        <svg class="w-6 h-6 text-grey-400"><use xlink:href="#icon-document"/></svg>
+                        <svg class="w-6 h-6 text-grey-400">
+                            <use xlink:href="#icon-document"/>
+                        </svg>
                     @endif
                 </div>
             </a>
@@ -27,20 +34,22 @@
             @if($count === 1)
                 <div class="space-y-0.5 leading-tight py-1.5 grow">
                     <p class="body-dark">
-                        {{ $file->getFileName() }}
+                        {{ $asset->getFileName() }}
                     </p>
 
                     <p class="text-sm text-grey-500">
-                        {{-- TODO: $file->isExternalAsset always return null --}}
-                        @if($file->isExternalAsset)
-                            {{ ucfirst($file->getExternalAssetType()) }} -
-                            {{ $file->getData('external.duration') }} sec
-                        @else
-                            {{ $file->humanReadableSize }} -
-                            @if($file->isImage())
-                                {{ $file->width }}x{{ $file->height }} -
+                        @if($asset->hasData('external'))
+                            {{ ucfirst($asset->getData('external.type')) }}
+
+                            @if($asset->hasData('external.duration'))
+                                - {{ $asset->getData('external.duration') }} sec
                             @endif
-                            {{ strtoupper($file->extension) }}
+                        @else
+                            {{ $asset->getHumanReadableSize() }} -
+                            @if($asset->isImage())
+                                {{ $asset->getWidth() }}x{{ $asset->getHeight() }} -
+                            @endif
+                            {{ strtoupper($asset->getExtension()) }}
                         @endif
                     </p>
                 </div>
@@ -48,7 +57,9 @@
         </div>
     @empty
         <div class="flex items-center justify-center overflow-hidden rounded-lg w-14 h-14 shrink-0 bg-grey-100">
-            <svg class="w-6 h-6 text-grey-400"><use xlink:href="#icon-document"/></svg>
+            <svg class="w-6 h-6 text-grey-400">
+                <use xlink:href="#icon-document"/>
+            </svg>
         </div>
     @endforelse
 </div>