+++ CHANGELOG.md
@@ -5,7 +5,8 @@ the [Keep a CHANGELOG](http://keepachangelog.com/)
 principles.
 
 ## Unreleased
--   Added: chiefuser as key in db relation morphMap to Chief User model.
+- Fixed: Some file metadata was missing in file window.  
+- Added: chiefuser as key in db relation morphMap to Chief User model.
 
 ## 0.8.14 - 2024-02-20
 -   Fixed: dependency livewire fixed to 3.4.4 because of rescan missing since 3.4.5