+++ CHANGELOG.md
@@ -7,6 +7,7 @@ principles.
 ## Unreleased
 - Fixed: Some file metadata was missing in file window.  
 - Added: chiefuser as key in db relation morphMap to Chief User model.
+- Added: Tags to eager loading on the index results
 
 ## 0.8.14 - 2024-02-20
 -   Fixed: dependency livewire fixed to 3.4.4 because of rescan missing since 3.4.5